Refactor into base background job
