+++ src/Promitor.Agents.ResourceDiscovery/Scheduling/AzureSubscriptionDiscoveryBackgroundJob.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using CronScheduler.Extensions.Scheduler;
@@ -7,45 +6,15 @@ using GuardNet;
 using Microsoft.Extensions.Logging;
 using Prometheus.Client;
 using Promitor.Agents.ResourceDiscovery.Graph.Model;
-using Promitor.Agents.ResourceDiscovery.Repositories;
 using Promitor.Agents.ResourceDiscovery.Repositories.Interfaces;
-using Promitor.Core.Metrics;
 
 namespace Promitor.Agents.ResourceDiscovery.Scheduling
 {
-    public class DiscoveryBackgroundJob
-    {
-        private readonly IMetricFactory _metricFactory;
-        protected IAzureResourceRepository AzureResourceRepository { get; }
-        protected ILogger Logger { get; }
-
-        public DiscoveryBackgroundJob(IAzureResourceRepository azureResourceRepository, IMetricFactory metricFactory, ILogger logger)
-        {
-            Guard.NotNull(metricFactory, nameof(metricFactory));
-            Guard.NotNull(azureResourceRepository, nameof(azureResourceRepository));
-
-            Logger = logger;
-            _metricFactory = metricFactory;
-            AzureResourceRepository = azureResourceRepository;
-        }
-
-        protected void WritePrometheusMetric(string metricName, string metricDescription, int value, Dictionary<string, string> labels)
-        {
-                        // Order labels alphabetically
-            var orderedLabels = labels.OrderByDescending(kvp => kvp.Key).ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
-
-            var gauge = _metricFactory.CreateGauge(metricName, help: metricDescription, includeTimestamp: true, labelNames: orderedLabels.Keys.ToArray());
-            gauge.WithLabels(orderedLabels.Values.ToArray()).Set(value);
-        }
-    }
     public class AzureSubscriptionDiscoveryBackgroundJob : DiscoveryBackgroundJob, IScheduledJob
     {
         public const string MetricName = "promitor_azure_landscape_subscription_info";
         public const string MetricDescription = "Provides information concerning the Azure subscriptions in the landscape that Promitor has access to.";
         
-        // TODO: Refactor this one
-        private readonly IRuntimeMetricsCollector _runtimeMetricsCollector;
-
         public AzureSubscriptionDiscoveryBackgroundJob(string jobName, IAzureResourceRepository azureResourceRepository,  IMetricFactory metricFactory, ILogger<AzureSubscriptionDiscoveryBackgroundJob> logger)
             : base(azureResourceRepository, metricFactory, logger)
         {
@@ -58,7 +27,7 @@ namespace Promitor.Agents.ResourceDiscovery.Scheduling
 
         public async Task ExecuteAsync(CancellationToken cancellationToken)
         {
-            Logger.LogInformation("Discovering Azure Landscape!");
+            Logger.LogTrace("Discovering Azure subscriptions...");
 
             // Discover Azure subscriptions
             var discoveredLandscape = await AzureResourceRepository.DiscoverAzureSubscriptionsAsync();
@@ -68,6 +37,8 @@ namespace Promitor.Agents.ResourceDiscovery.Scheduling
             {
                 ReportDiscoveredAzureInfo(discoveredLandscapeItem);
             }
+
+            Logger.LogTrace("Azure subscriptions discovered...");
         }
 
         private void ReportDiscoveredAzureInfo(AzureSubscriptionInformation azureLandscapeInformation)
